script for copying/syncing app preferences

+++ preferences.sh
@@ -1,13 +1,22 @@
 # Sync preference files
 DROPBOX="$HOME/Dropbox/app settings"
-PREFERENCES="~/Library/Preferences"
+PREFERENCES="$HOME/Library/Preferences"
 
 DIVVY="com.mizage.Divvy.plist"
 
-# Create symlinks from preferences to dropbox
-cd "$DROPBOX"
-ln -s "$PREFERENCES/$DIVVY"
-
-# Create symlinks from dropbox to preferences
-cd $PREFERENCES
-ln -s "$DROPBOX/$DIVVY"
+if  [[ $1 = "setup" ]]; then
+    echo "Copying your preference files to Dropbox"
+    cp "$PREFERENCES/$DIVVY" $DROPBOX
+elif [[ $1 = "sync" ]]; then
+    echo "Are you sure you want to sync? This will override your settings (yes/no): "
+    read confirmation
+    if  [[ $confirmation != "yes" ]]; then
+        exit
+    fi
+    echo "Copying your preference files from Dropbox"
+    cd $PREFERENCES
+    rm $DIVVY
+    ln -s "$DROPBOX/$DIVVY"
+else
+    echo 'Invalid option'
+fi
